fix code duplication in funtofem_callback and imp readability
