+++ pyfuntofem/interface/utils/__init__.py
@@ -11,6 +11,11 @@ fun3d_loader = importlib.util.find_spec("fun3d")
 from .cart3d_utils import *
 from .loader_decorators import *
 
+# tacs utils
+tacs_loader = importlib.utils.find_spec("tacs")
+if tacs_loader is not None:
+    from .funtofem_callback import *
+
 # need to be updated and therefore commmented out for now
 # if openmdao_loader is not None: from .openmdao import *
 # if fun3d_loader is not None: from .fun3d_client import *
